+ update to BQv3
