+++ build.gradle
@@ -17,12 +17,12 @@ buildscript {
 
 apply plugin: 'forge'
 
-version = "1"
-group= "lokko12.CB4BQ" // http://maven.apache.org/guides/mini/guide-naming-conventions.html
+version = "3"
+group= "com.github.bartimaeusnek.CB4BQ" // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 archivesBaseName = "CB4BQ"
 
 minecraft {
-    version = "1.7.10-10.13.4.1558-1.7.10"
+    version = "1.7.10-10.13.4.1614-1.7.10"
     runDir = "eclipse"
 }
