+++ src/main/java/com/github/bartimaeusnek/CB4BQ/ServerProxy.java
@@ -1,10 +1,7 @@
-package lokko12.CB4BQ;
+package com.github.bartimaeusnek.CB4BQ;
 
 import cpw.mods.fml.common.event.FMLInitializationEvent;
-import cpw.mods.fml.common.event.FMLPostInitializationEvent;
-import cpw.mods.fml.common.event.FMLPreInitializationEvent;
 import cpw.mods.fml.common.registry.GameRegistry;
-import net.minecraft.block.Block;
 
 public class ServerProxy {
